Max Zoom of layer on map type change
This little fix prevents that the google terrain map is out of synch with the overlays if the leaflet zoom level is too high.

The check only works when you switch to the terrain layer. It would be nice to find a way to prevent the user to zoom too far on the terrain layer too.