L.Yandex: do not try to override controls panel margins
It's bad practice, and may conflict with other leaflet controls, css, and actual developer intentions.
And it even do not work in some cases, e.g. when 2 yandex layers added, and then remove 1.

Besides, we do not need two attribution controls anyway (and it looks ugly).
It'd be better to gather all info in 1 (either Yandex's, either Leaflet's), and there is API for this.

Choosing appropriate decision is up on developer now.
If he want to change margins - let him do this explicitely.
And there are plenty of other options, among them:
- move attribution control to other corner
- disable attribution control completely
- implement dynamic attribution fetching from one control to another