L.Yandex: fix z-order of overlay
Previously overlay was positioned over map pane, and this prevented normal
interaction with features (like markers, popups, etc).

Correct z-order can be provided only if overlay belongs to same stacking context
as the rest of panes, thats why we have to make it descendant of leaflet-map-pane.

But as map pane is panning on drag, it's now necessary to update ymaps offset in opposite order
to fit into visible window.

---
Note, that currently ymaps controls/copyrights belong to the rest of layer:
- positioned *behind* leaflet controls / features
- moving along with ground (tile) layer on drag
  (not an issue when updating is fast)

It's possible to fix it, if we divide ymaps panes into different containers.
That could lead to other issues thought..
So let it as is for now.